% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_element_kahan}
\alias{add_element_kahan}
\title{Numerically safe addition using Kahan summation}
\usage{
add_element_kahan(sum, element, compensation)
}
\arguments{
\item{&sum}{Current accumulated sum. Updated by function.}

\item{&element}{Element to add to the accumulated sum. Not updated.}

\item{&compensation}{Current adjustment to compensate for floating point
summation error. Updated by function.}
}
\description{
Uses Kahan summation algorithm to limit numerical error caused by adding
lots of very small things to a big thing to O(1) rather than O(N)
See: https://en.wikipedia.org/wiki/Kahan_summation_algorithm

Note that particularly agressive compiler optimisations can result in the
Kahan compensation being optimised away. We believe that this requires the
-ffast-math compiler flag to be explicitly set, and we attempt to ensure it
to override any local setting for this flag by adding the -fno-fast-math 
flag to PKG_CPPFLAGS in src/Makevars.
}
