% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{area_between_dhist_ecmfs}
\alias{area_between_dhist_ecmfs}
\title{Calculate area between two discrete histogram empirical cumulative 
mass functions (ECMFs)}
\usage{
area_between_dhist_ecmfs(dhist_ecmf1, dhist_ecmf2)
}
\arguments{
\item{dhist_ecmf1}{An object of class \code{dhist_ecmf}, returned from a call 
to the \code{dhist_ecmf} function}

\item{dhist_ecmf2}{An object of class \code{dhist_ecmf}, returned from a call 
to the \code{dhist_ecmf} function}
}
\value{
area The area between the two discrete histogram ECMFs, calculated as
the integral of the absolute difference between the two ECMFs
}
\description{
Calculate area between two discrete histogram empirical cumulative 
mass functions (ECMFs)
}
