% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{as_smoothed_dhist}
\alias{as_smoothed_dhist}
\title{Set dhist smoothing}
\usage{
as_smoothed_dhist(dhist, smoothing_window_width)
}
\arguments{
\item{smoothing_window_width}{If greater than 0, the discrete histogram will
be treated as having the mass at each location "smoothed" uniformly across
a bin centred on the location and having width = \code{smoothing_window_width}}
}
\value{
A copy of a \code{dhist} object with its \code{smoothing_window_width}
attribute set  to the value provided \code{smoothing_window_width} parameter.
}
\description{
Returns a "smoothed" copy of a \code{dhist} object with its 
\code{smoothing_window_width} attribute set to the value provided 
\code{smoothing_window_width} parameter.
}
