% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{as_unsmoothed_dhist}
\alias{as_unsmoothed_dhist}
\title{Remove dhist smoothing}
\usage{
as_unsmoothed_dhist(dhist)
}
\value{
A copy of a \code{dhist} object with its \code{smoothing_window_width}
attribute set to 0.
}
\description{
Returns an "unsmoothed" copy of a \code{dhist} object with its 
\code{smoothing_window_width} attribute set to 0.
}
