% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-net-emd.R
\name{augment_histograms}
\alias{augment_histograms}
\title{Augment a pair of histograms to share a common set of bins}
\usage{
augment_histograms(bin_masses1, bin_masses2, bin_centres1, bin_centres2)
}
\arguments{
\item{bin_masses1}{Bin masses for histogram 1}

\item{bin_masses2}{Bin masses for histogram 2}

\item{bin_centres1}{Bin centres for histogram 1}

\item{bin_centres2}{Bin centres for histogram 2}
}
\description{
Where a bin centre only exists in one histogram, add a zero mass bin with the
same centre to the other histogram to ensure that all bins exist in both 
histograms. This makes some hsitogram operations easier.
}
