% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_binning.R
\name{binned_densities_adaptive}
\alias{binned_densities_adaptive}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
binned_densities_adaptive(densities, min_counts_per_interval,
  num_intervals)
}
\description{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to 
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
Temporarily accessible during development. 
TODO: Remove @export prior to publishing
}
