% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-net-emd.R
\name{cost_matrix}
\alias{cost_matrix}
\title{Generates a matrix for the cost of moving a unit of mass between all pairs of
features. The cost is the distance between the locations of each feature pair}
\usage{
cost_matrix(bin_locations1, bin_locations2)
}
\arguments{
\item{bin_locations1}{Feature bin locations for network 1}

\item{bin_locations2}{Feature bin locations for network 2}
}
\description{
Generates a matrix for the cost of moving a unit of mass between all pairs of
features. The cost is the distance between the locations of each feature pair
}
