% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{count_graphlets_ego_scaled}
\alias{count_graphlets_ego_scaled}
\title{Scaled graphlet count for ego-networks}
\usage{
count_graphlets_ego_scaled(graph, max_graphlet_size, neighbourhood_size,
  min_ego_nodes = 3, min_ego_edges = 1, return_ego_networks = FALSE)
}
\arguments{
\item{graph}{A connected, undirected, simple graph as an \code{igraph} object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}

\item{neighbourhood_size}{The number of steps from the source node to include
nodes for each ego-network.}

\item{min_ego_nodes}{Only ego networks with at least \code{min_ego_nodes} 
nodes are returned.}

\item{min_ego_edges}{Only ego networks with at least \code{min_ego_edges} 
edges are returned.}

\item{return_ego_networks}{If \code{TRUE}, return ego-networks alongside 
graphlet counts to enable further processing.}
}
\value{
If \code{return_ego_networks = FALSE}, returns an RxC matrix 
containing counts of each graphlet (columns, C) for each ego-network in the 
input graph (rows, R). Columns are labelled with graphlet IDs and rows are 
labelled with the ID of the central node in each ego-network (if nodes in the
input graph are labelled). If \code{return_ego_networks = TRUE}, returns a
list with the following elements:
\itemize{
  \item \code{graphlet_counts}: A matrix containing graphlet counts for each 
  ego-network in the input graph as described above.
  \item \code{ego_networks}: The ego-networks of the query graph.
}
}
\description{
Calculates graphlet counts for the n-step ego-network of each node in a graph, 
scaled by dividing the graphlet counts for each ego-network by the total
number of possible groupings of nodes in the ego-network with the same number
of nodes as each graphlet. This scaling factor is choose(n, k), where n is the
number of nodes in the ego-network and k is the number of nodes in the graphlet.
}
