% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{count_graphlets_for_graph}
\alias{count_graphlets_for_graph}
\title{Count total number of graphlets in a graph}
\usage{
count_graphlets_for_graph(graph, max_graphlet_size)
}
\arguments{
\item{graph}{A connected, undirected, simple graph as an \code{igraph} object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
Vector containing counts of each graphlet for the graph.
}
\description{
Calculates total graphlet counts for a \code{igraph} graph object using the
ORCA fast graphlet orbit counting package. Per-node graphlet counts are
calculated by summing orbits over graphlets. These are then divided by the
number of nodes comprising each graphlet to avoid counting the same graphlet
multiple times.
}
