% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{count_orbits}
\alias{count_orbits}
\title{Calculate graphlet orbit counts}
\usage{
count_orbits(graph, max_graphlet_size)
}
\arguments{
\item{graph}{An \code{igraph} graph object. The graph must have the following
properties:' 
  1. Be undirected
  2. Have no self-loops (where both endpoints of an edge are the same vertex)
  3. Not have multiple edges multiple edges (i.e. at most one edge exists for
     each pair of vertices)
  4. No isolated vertices (i.e. vertices with no edges)}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
ORCA-format matrix containing counts of each graphlet
orbit (columns) at each vertex in the graph (rows).
}
\description{
Calculates graphlet orbit counts from \code{igraph} graph object,
using the ORCA fast graphlet orbit counting package.
}
