% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{counts_from_observations}
\alias{counts_from_observations}
\title{Count number of occurences}
\usage{
counts_from_observations(features)
}
\arguments{
\item{features}{A Matrix with doubles.}
}
\description{
Count number of occurences
}
