% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{cross_comp_to_matrix}
\alias{cross_comp_to_matrix}
\title{Convert a pair-wise cross-comparison into a matrix format}
\usage{
cross_comp_to_matrix(measure, cross_comparison_spec)
}
\arguments{
\item{measure}{A list of pair-wise comparison measiures}

\item{cross_comparison_spec}{A cross-comparison specification generated
using \code{cross_comparison_spec}}
}
\value{
A square symmetric matrix with a zero diagonal, with elements
Cij and Cji populated from the element from \code{measure} corresponding to
the row of \code{cross_comparison_spec} with \code{index_a = i} and 
\code{index_b = j}
}
\description{
Converts a pair-wise cross-comparison into a matrix format
}
