% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{density_binned_counts}
\alias{density_binned_counts}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
density_binned_counts(graphlet_counts, density_interval_indexes,
  agg_fn = mean, scale_fn = NULL, ego_networks = NULL,
  max_graphlet_size = NULL)
}
\arguments{
\item{agg_fn}{Function to aggregate counts in each bin (default \code{agg_fn = mean}).}

\item{scale_fn}{Optional function to apply a transformation to graphlet_counts, must
have arguments graphlet_counts, ego_networks and max_graphlet_size.}

\item{ego_networks}{Optionally passed and used by scale_fn.}

\item{max_graphlet_size}{Optionally passed and used by scale_fn.
Temporarily accessible during development.
TODO: Remove @export prior to publishing}
}
\description{
Used to
generate a function for calculating expected graphlet counts in each
density bin.
}
