% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{density_binned_counts_gp}
\alias{density_binned_counts_gp}
\title{Calculate expected counts in density bins using geometric poisson (Polya-Aeppli) approximation}
\usage{
density_binned_counts_gp(graphlet_counts, density_interval_indexes,
  max_graphlet_size)
}
\description{
Calculate expected counts in density bins using geometric poisson (Polya-Aeppli) approximation
}
