% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{density_from_counts}
\alias{density_from_counts}
\title{Calculate edge density for a single graph.}
\usage{
density_from_counts(graphlet_counts)
}
\arguments{
\item{graphlet_counts}{Vector of pre-calculated graphlet, edge and node
counts. Must have named items "N" (node counts) and "G0" (edge counts).}
}
\description{
Calculate edge density for a single graph.
}
