% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{dhist_from_obsSLOW}
\alias{dhist_from_obsSLOW}
\title{Discrete histogram from observations}
\usage{
dhist_from_obsSLOW(observations)
}
\arguments{
\item{observations}{A vector of discrete numeric observations}
}
\value{
A sparse discrete histogram. Format is a \code{dhist} object, which
is a list of class \code{dhist} with the following named elements:
\itemize{
  \item \code{locations}: A 1D numeric vector of discrete locations
  \item \code{masses}: A 1D numeric vector of the mass present at each location
}
}
\description{
Generate a sparse discrete histogram from a set of discrete numeric observations
}
