% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{dhist_std}
\alias{dhist_std}
\title{Calculate standard deviation of a discrete histogram}
\usage{
dhist_std(dhist)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}
}
\value{
Standard deviation of histogram
}
\description{
Calculates standard deviation directly from the discrete histogram by using
locations weighted by masses.
NOTE: Does not apply bias correction (i.e. N-1 denominator) as masses
may not represent counts so N is not necessarily known
}
