% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{dhist_variance}
\alias{dhist_variance}
\title{Calculate variance of a discrete histogram}
\usage{
dhist_variance(dhist)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}
}
\value{
Variance of histogram
}
\description{
Calculates variance directly from the discrete histogram by using locations
weighted by  masses. 
NOTE: Does not apply bias correction (i.e. N-1 denominator) as masses 
may not represent counts so N is not necessarily known
}
