% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{discrete_hist}
\alias{discrete_hist}
\title{Discrete histogram}
\usage{
discrete_hist(observations)
}
\arguments{
\item{observations}{A vector of integer-valued observations}
}
\value{
A sparse discrete histogram. Format is a list with the following 
named elements:
\itemize{
  \item \code{$locations}: Integer-valued locations where one or more observations 
  exist
  \item \code{$masses}: Integer counts of observations at each location
}
}
\description{
Generate a sparse discrete histogram from a set of integer observations
}
