% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{ego_network_node_counts}
\alias{ego_network_node_counts}
\title{ego_network_node_counts}
\usage{
ego_network_node_counts(ego_networks)
}
\arguments{
\item{ego_networks}{Named list of ego networks for a graph.}
}
\description{
Calculates number of nodes in each ego network.
}
