% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{ego_to_graphlet_counts}
\alias{ego_to_graphlet_counts}
\title{ego_to_graphlet_counts
JACK To follow through logic of paper steps, wanted to pass
ego networks to the function for generating graphlet counts,
not the input query graph directly (as in count_graphlets_ego above).}
\usage{
ego_to_graphlet_counts(ego_networks, max_graphlet_size = 4)
}
\arguments{
\item{ego_networks}{Named list of ego networks for a graph.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
returns an RxC matrix
containing counts of each graphlet (columns, C) for each ego-network (rows, R).
Columns are labelled with graphlet IDs and rows are
labelled with the ID of the central node in each ego-network.
}
\description{
Calculates graphlet counts for previously generated ego networks.
}
