% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd}
\alias{emd}
\title{Earth Mover's Distance (EMD)}
\usage{
emd(dhist1, dhist2, smoothing_window_width = 0, normalise_mass = FALSE,
  normalise_variance = FALSE)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}

\item{smoothing_window_width}{Width of "top-hat" smoothing window to apply to
"smear" point masses across a finite width in the real domain. Default is 0, 
which  results in no smoothing. Care should be taken to select a 
\code{smoothing_window_width} that is appropriate for the discrete domain 
(e.g.for the integer domain a width of 1 is the natural choice)}

\item{normalise_mass}{Logical determining whether histograms are normalised 
to  unit mass prior to calculating the EMD (default = FALSE)}

\item{normalise_variance}{Logical determining whether histograms are normalised 
to  unit variance prior to calculating the EMD (default = FALSE)}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the Earth Mover's Distance (EMD) between two discrete histograms
}
