% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd_cs}
\alias{emd_cs}
\title{Earth Mover's Distance (EMD) using the difference of cumulative sums method}
\usage{
emd_cs(dhist1, dhist2, smoothing_window_width = 0, normalise_mass = FALSE,
  normalise_variance = FALSE)
}
\arguments{
\item{dhist1}{A discrete histogram as a \code{dhist} object}

\item{dhist2}{A discrete histogram as a \code{dhist} object}

\item{smoothing_window_width}{Width of "top-hat" smoothing window to apply to
"smear" point masses across a finite width in the real domain. Default is 0, 
which  results in no smoothing. Care should be taken to select a 
\code{smoothing_window_width} that is appropriate for the discrete domain 
(e.g.for the integer domain a width of 1 is the natural choice)}

\item{normalise_mass}{Logical determining whether histograms are normalised 
to  unit mass prior to calculating the EMD (default = FALSE)}

\item{normalise_variance}{Logical determining whether histograms are normalised 
to  unit variance prior to calculating the EMD (default = FALSE)}
}
\value{
Earth Mover's Distance between the two input histograms
}
\description{
Takes two discrete histograms and calculates the Wasserstein / Earth Mover's
Distance between the two histograms by summing the absolute difference 
between the two cumulative histograms.
}
\references{
Calculation of the Wasserstein Distance Between Probability Distributions on the Line
S. S. Vallender, Theory of Probability & Its Applications 1974 18:4, 784-786
\url{http://dx.doi.org/10.1137/1118101}
}
