% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emd_fast_no_smoothing}
\alias{emd_fast_no_smoothing}
\title{Compute Earth Mover's Distance (EMD) between two Empirical Cumulative
Density Functions (ECDFs)}
\usage{
emd_fast_no_smoothing(locations1, values1, locations2, values2)
}
\arguments{
\item{locations1}{Locations for ECDF 1}

\item{values1}{Cumulative masses for ECDF 1}

\item{locations2}{Locations for ECDF 2}

\item{values2}{Cumulative masses for ECDF 2}
}
\description{
Compute Earth Mover's Distance (EMD) between two Empirical Cumulative
Density Functions (ECDFs)
}
