% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures-net-emd.R
\name{emd_lp}
\alias{emd_lp}
\title{Earth Mover's Distance (EMD) using linear programming (LP)
Takes two sets of feature counts and locations and calculates the 
Earth Mover's Distance between them. EMD is also known as the Wasserstein 
distance or metric
Required parameters are the counts of each feature for both networks}
\usage{
emd_lp(bin_counts1, bin_counts2, bin_locations1, bin_locations2)
}
\arguments{
\item{bin_counts1}{Feature counts for network 1}

\item{bin_counts2}{Feature counts for network 2
Optional parameters are the locations of each feature. If omitted, it is 
assumed that features are separated by unit distance}

\item{bin_locations1}{Feature bin locations for network 1}

\item{bin_locations2}{Feature bin locations for network 2}
}
\description{
Earth Mover's Distance (EMD) using linear programming (LP)
Takes two sets of feature counts and locations and calculates the 
Earth Mover's Distance between them. EMD is also known as the Wasserstein 
distance or metric
Required parameters are the counts of each feature for both networks
}
