% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{==.dhist}
\alias{==.dhist}
\title{Compare dhists}
\usage{
\method{==}{dhist}(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A discrete histogram as a \code{dhist} object}

\item{dhist2}{A discrete histogram as a \code{dhist} object}
}
\description{
Compares all fields of the dhist and only returns treu if they are all the
same in both dhists
}
