% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{exp_counts_bin_gp}
\alias{exp_counts_bin_gp}
\title{INTERNAL FUNCTION - DO NOT CALL DIRECTLY
Used by \code{density_binned_counts_gp}
Calculate expected counts with geometric poisson (Polya-Aeppli)
approximation for a single density bin.}
\usage{
exp_counts_bin_gp(bin_idx, graphlet_counts, density_interval_indexes,
  max_graphlet_size)
}
\arguments{
\item{bin_idx}{Density bin index to calculate expected counts for.}

\item{graphlet_counts}{Graphlet counts for a number of ego_networks.}

\item{density_interval_indexes}{Density bin indexes for each ego network in
\code{graphlet_counts}.}

\item{max_graphlet_size}{Determines the maximum size of graphlets
included in graphlet_counts.}
}
\description{
INTERNAL FUNCTION - DO NOT CALL DIRECTLY
Used by \code{density_binned_counts_gp}
Calculate expected counts with geometric poisson (Polya-Aeppli)
approximation for a single density bin.
}
