% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{gdd}
\alias{gdd}
\title{Graphlet-based Degree Distributions (GDDs)}
\usage{
gdd(indexed_edges, type = "node4")
}
\arguments{
\item{indexed_edges}{A 2 x numEdges edgelist with vertices labelled with 
integer indices, with an optional "vertex_names" attribute}

\item{type}{Type of graphlet-based degree distributions: "orb4" counts all 
orbits for graphlets comprising up to 4 nodes; "orb5" counts all orbits for
graphlets comprising up to 5 nodes.}
}
\value{
List of graphlet-based degree distributions, with each distribution
represented as a \code{dhist} discrete histogram object.
}
\description{
Calculates the specified set of graphlet-based degree distributions from an 
indexed edge list using the ORCA fast graphlet orbit counting package.
}
