% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{gdd_for_all_graphs}
\alias{gdd_for_all_graphs}
\title{Load all graphs in a directory and calculates their Graphlet-based Degree
Distributions (GDDs)}
\usage{
gdd_for_all_graphs(source_dir, format = "ncol", pattern = ".txt",
  feature_type = "orbit", max_graphlet_size = 4,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{source_dir}{Path to graph directory}

\item{format}{Format of graph files}

\item{pattern}{Filename pattern to match graph files}

\item{type}{Type of graphlet-based degree distributions: "orb4" counts all 
orbits for graphlets comprising up to 4 nodes; "orb5" counts all orbits for
graphlets comprising up to 5 nodes.}
}
\value{
A named list where each element contains a set of GDDs for a single 
graph from the source directory. Each set of GDDs is itself a named list,  
where each GDD element is a \code{dhist} discrete histogram object.
}
\description{
Loads graphs from all files matching the given pattern in the given directory,
converts them to indexed edge lists compatible with the ORCA fast orbit 
counting package and calculates the specified set of graphlet-based degree 
distributions usingthe ORCA package.
}
