% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_comparison.R
\name{graph_cross_comparison_spec}
\alias{graph_cross_comparison_spec}
\title{Generate graph cross-comparison specification}
\usage{
graph_cross_comparison_spec(graph_edges)
}
\arguments{
\item{graph_edges}{List of graphs in ORCA-compatible graph edge format}
}
\value{
M
}
\description{
Generate graph cross-comparison specification
}
