% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_graph_functions.R
\name{graph_to_indexed_edge_list}
\alias{graph_to_indexed_edge_list}
\title{Integer index edge list from igraph}
\usage{
graph_to_indexed_edge_list(graph)
}
\arguments{
\item{graph}{An igraph graph object}
}
\value{
A 2 x numEdges edgelist with vertices labelled with integer indices
}
\description{
Takes a igraph graph object and generates an edgelist where each edge is
represented by the integer indexes of its vertices. Note that, where a graph
has isolated vertices, the indexes for these vertices will not be present
in the edge list. Where a graph has no isolated vertices, the edge list will 
include all vertex indexes from 1 to numVertices.
}
