% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{graph_to_orca_graph}
\alias{graph_to_orca_graph}
\title{Make arbitrary igraph ORCA compatible}
\usage{
graph_to_orca_graph(graph)
}
\arguments{
\item{graph}{Original igraph graph object}
}
\value{
An ORCA compatible igraph graph object
}
\description{
Takes a igraph graph object and does the following to make the graph
connected, undirected and simple to ensure compatiblity with the ORCA
fast orbit calculation package:
  1. Makes the graph undirected
  2. Removes loops (where both endpoints of an edge are the same vertex)
  3. Removes multiple edges (i.e. ensuring only one edge exists for each 
     pair of endpoints)
  4. Removes isolated vertices (i.e. vertices with no edges after the 
     previous alterations)
}
