% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{graphlet_ids_for_size}
\alias{graphlet_ids_for_size}
\title{Graphlet IDs for size}
\usage{
graphlet_ids_for_size(graphlet_size)
}
\arguments{
\item{graphlet_size}{Graphlet size (i.e. number of nodes in the graphlet)}
}
\value{
A vector containing the IDs of all graphlets made up of the specified
number of nodes
}
\description{
List IDs for all graphlets of a specified size
}
