% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{graphlet_key}
\alias{graphlet_key}
\title{Graphlet key}
\usage{
graphlet_key(max_graphlet_size)
}
\arguments{
\item{max_graphlet_size}{Maximum number of nodes graphlets can contain}
}
\value{
Metadata list with the following named fields:
\itemize{
  \item \code{max_nodes}: Maximum number of nodes graphlets can contain
  \item \code{id}: ID of each graphlet in format Gn, where n is in range 0 to
 num_graphlets
  \item \code{node_count}: Number of nodes contained within each graphlet
}
}
\description{
Metdata about graphlet groups.
}
