% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{harmonise_dhist_locations}
\alias{harmonise_dhist_locations}
\title{Harmonise a pair of discrete histograms to share a common set of locations}
\usage{
harmonise_dhist_locations(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A discrete histogram as a \code{dhist} object}

\item{dhist2}{A discrete histogram as a \code{dhist} object}
}
\value{
Harmonised histograms
}
\description{
Where a location only exists in one histogram, add this location to the other
histogram with zero mass. This ensures that all location exist in both 
histograms.
}
