% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{histogram_variance}
\alias{histogram_variance}
\title{Calculate histogram variance}
\usage{
histogram_variance(bin_masses, bin_centres)
}
\arguments{
\item{bin_masses}{Bin masses for histogram}

\item{bin_centres}{Bin centres for histogram}
}
\value{
Variance of histogram
}
\description{
Calculates variance directly from the histogram by using bin centres weighted
by bin masses. Where a histogram represents a quantised summary of an 
underlying sample, it is more accurate to calculate variance directly from 
the sample data if available.
NOTE: Does not apply bias correction (i.e. N-1 denominator) as bin_masses 
may not represent bin counts so N is not necessarily known
}

