% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_graph_functions.R
\name{indexed_edges_to_graph}
\alias{indexed_edges_to_graph}
\title{Graph from integer index edge list}
\usage{
indexed_edges_to_graph(indexed_edges)
}
\arguments{
\item{indexed_edges}{A 2 x numEdges edgelist with vertices labelled with 
integer indices, with an optional "vertex_names" attribute}
}
\value{
An igraph graph object
}
\description{
Takes an integer indexed edgelist (where each edge is represented by the 
integer indexes of its vertices) and converts it to an igraph format graph.
If the edge list has a "vertex_names" attribute, this will be used to name
the vertices in the resultant graph.
}
