% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_graph_functions.R
\name{integer_index_edge_list}
\alias{integer_index_edge_list}
\title{Integer index edge list from igraph}
\usage{
integer_index_edge_list(graph)
}
\arguments{
\item{graph}{igraph graph object}
}
\value{
edges A 2 x numEdges edgelist with integer vertex indices
}
\description{
Takes a igraph graph object and generates an edgelist where each edge is
represented by the integer IDs of its vertices. Note that, where a graph
has isolated vertices, the IDs for these vertices will not be present
in the edge list. Where a graph has no isolated vertices, the edge list will 
include all vertex IDs from 1 to numVertices.
}
