% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{interval_index}
\alias{interval_index}
\title{Bin values into intervals based on the provided breaks}
\usage{
interval_index(x, breaks, out_of_range_intervals = FALSE)
}
\arguments{
\item{x}{The values to be binned}

\item{breaks}{The boundaries between bins}

\item{out_of_range_intervals}{If \code{TRUE}, "out of range" values lying
below the first break or above the last break will be assigned to additional 
unbounded lower and upper extrema intervals. If \code{FALSE} these "out of 
range" values will be assigned to intervals bounded by the lowest or 
uppermost break respectively.}
}
\value{
A vector of bin indexes, one per value provided
}
\description{
Bin values into intervals based on the provided breaks
}
