% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{is_dhist}
\alias{is_dhist}
\title{Check if an object is a \code{dhist} discrete histogram}
\usage{
is_dhist(x, fast_check = TRUE)
}
\arguments{
\item{x}{An arbitrary object}

\item{fast_check}{Boolean flag indicating whether to perform only a
superficial fast check limited to checking the object's class attribute
is set to \code{dhist} (default = \code{TRUE})}
}
\description{
Checks if the input object is of class \code{dhist}. If \code{fast_check} is
\code{TRUE} then the only check is whether the object has a class attribute
of \code{dhist}. If \code{fast_check} is \code{FALSE} (default), then checks
are also made to ensure that the object has the structure required of a
\code{dhist} object.
}
