% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{is_numeric_vector_1d}
\alias{is_numeric_vector_1d}
\title{Check if 1D numeric vector}
\usage{
is_numeric_vector_1d(input)
}
\arguments{
\item{input}{Arbitrary object}
}
\value{
TRUE if input is a 1D numeric vector. FALSE otherwise.
}
\description{
Check if a variable is a 1D numeric vector by checking that:
\itemize{
  \item \code{is_numeric(input)}: Input is vector, matrix, array or list of numbers
  \item \code{is_null(dim(input))}: Input is not a matrix or array
}
}
