% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{make_named_ego_graph}
\alias{make_named_ego_graph}
\title{Get ego-networks for a graph as a named list}
\usage{
make_named_ego_graph(graph, order, ...)
}
\arguments{
\item{graph}{An \code{igraph} object}

\item{order}{The number of steps from the source node to include
nodes for each ego-network.}

\item{...}{Additional parameters to be passed to the underlying 
\code{igraph::make_ego_graph} function used.}
}
\description{
Simple wrapper for the \code{igraph::make_ego_graph} function that names
each ego-network in the returned list with the name of the node in the
original graph that the ego-network was generated from
}
