% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{mean_centre_dhist}
\alias{mean_centre_dhist}
\title{Centre a discrete histogram around its mean location}
\usage{
mean_centre_dhist(dhist)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}
}
\value{
The mass-weighted mean location
}
\description{
Centres a discrete histogram around its mass-weighted mean location by
subtracting the mass-weighted mean from each location.
}
