% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{mean_density_binned_graphlet_counts}
\alias{mean_density_binned_graphlet_counts}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
mean_density_binned_graphlet_counts(graphlet_counts,
  density_interval_indexes, agg_fn = mean)
}
\description{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
Temporarily accessible during development.
TODO: Remove @export prior to publishing
}
