% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{min_emd}
\alias{min_emd}
\title{Minimum Earth Mover's Distance (EMD)}
\usage{
min_emd(dhist1, dhist2, method = "optimise", ...)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}

\item{method}{The method to use to find the minimum EMD across all potential 
offsets for each pair of histograms. Default is "optimise".
\itemize {
 \item \code{optimise}: Default. Uses an optimiser to efficiently find the offset 
 with the minimal EMD. However, this is not guaranteed to find the global 
 minimum if multiple local minima EMDs exist. See \code{min_emd_optimise}
 for any method-specific arguments.
 \item Exhaustively evaluate the EMD between the histograms at all offsets that
 are candidates for the minimal EMD.See \code{min_emd_exhaustive}
 for any method-specific arguments.
 \item Evaluate the EMD between the histograms at a series of offsets separated
 by a fixed step and spanning the range of overalap of the two histograms. See 
 \code{min_emd_fixed_step} for any method-specific arguments.
}}

\item{...}{Method-specific arguments to be passed to the specific
method used to calculate the minimum EMD}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the minimum Earth Mover's Distance (EMD) between two discrete 
histograms. This is the minimum EMD between the two histograms across all
possible offsets of histogram 1 against histogram 2.
}
