% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{min_emd_exhaustive}
\alias{min_emd_exhaustive}
\title{Minimum Earth Mover's Distance (EMD) using exhaustive search}
\usage{
min_emd_exhaustive(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the minimum Earth Mover's Distance (EMD) between two discrete 
histograms using an exhaustive search.
}
\details{
When "sliding" two piecewise-linear empirical cumulative mass functions 
(ECMFs) across each other to minimise the EMD between them, it is sufficient 
to calculate the EMD at all offsets where any knots from the two ECMFs align 
to ensure that the offset with the global minimum EMD is found.

This is because of the piece-wise linear nature of the two ECMFs. Between any 
two offsets where knots from the two ECMFs align, EMD will be either constant,
or uniformly increasing or decreasing. Therefore, there the EMD between two 
sets of aligned knots cannot be smaller than the EMD at one or other of the 
bounding offsets.
}
