% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{min_emd_optimise}
\alias{min_emd_optimise}
\title{Minimum Earth Mover's Distance (EMD) using optimiser search}
\usage{
min_emd_optimise(dhist1, dhist2, tolerance = .Machine$double.eps * 1000)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}

\item{tol}{Tolerance that defines stopping condition for optimiser. See 
documentation for \code{\link{stats:optimise}} for details on its effect.}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the minimum Earth Mover's Distance (EMD) between two discrete 
histograms by minimising the offset parameter of the \code{emd} function 
using the built-in \code{stats::optimise} method.
}
