% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{min_emd_optimise_fast}
\alias{min_emd_optimise_fast}
\title{Minimum Earth Mover's Distance (EMD) using fast optimiser search}
\usage{
min_emd_optimise_fast(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the minimum Earth Mover's Distance (EMD) between two discrete 
histograms by minimising the offset parameter of the \code{emd} function 
using the built-in \code{stats::optimise} method.
}
