% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{net_emd_one_to_one}
\alias{net_emd_one_to_one}
\title{NetEMD Network Earth Mover's Distance between a pair of networks.}
\usage{
net_emd_one_to_one(
  graph_1 = NULL,
  graph_2 = NULL,
  dhists_1 = NULL,
  dhists_2 = NULL,
  method = "optimise",
  return_details = FALSE,
  smoothing_window_width = 0,
  feature_type = "orbit",
  max_graphlet_size = 5,
  ego_neighbourhood_size = 0
)
}
\arguments{
\item{graph_1}{A network/graph object from the \code{igraph} package. \code{graph_1} can be set to \code{NULL} (default) if dhists_1 is provided.}

\item{graph_2}{A network/graph object from the \code{igraph} package. \code{graph_2} can be set to \code{NULL} (default) if dhists_2 is provided.}

\item{dhists_1}{A \code{dhist} discrete histogram object or a list of such objects. \code{dhists_1} can be set to \code{NULL} (default) if \code{graph_1} is provided.}

\item{dhists_2}{A \code{dhist} discrete histogram object or a list of such objects.  \code{dhists_2} can be set to \code{NULL} (default) if \code{graph_2} is provided.}

\item{method}{The method to use to find the minimum EMD across all potential
offsets for each pair of histograms. Default is "optimise" to use
R's built-in \code{stats::optimise} method to efficiently find the offset
with the minimal EMD. However, this is not guaranteed to find the global
minimum if multiple local minima EMDs exist. You can alternatively specify the
"exhaustive" method, which will exhaustively evaluate the EMD between the
histograms at all offsets that are candidates for the minimal EMD at the cost of computational time.}

\item{return_details}{Logical indicating whether to return the individual
minimal EMDs and associated offsets for all pairs of histograms}

\item{smoothing_window_width}{Width of "top-hat" smoothing window to apply to
"smear" point masses across a finite width in the real domain. Default is 0,
which  results in no smoothing. Care should be taken to select a
\code{smoothing_window_width} that is appropriate for the discrete domain
(e.g.for the integer domain a width of 1 is the natural choice)}

\item{feature_type}{Type of graphlet-based feature to count: "graphlet"
counts the number of graphlets each node participates in; "orbit" calculates
the number of graphlet orbits each node participates in.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes will be
counted. Possible values are 3,4, and 5 (default).}

\item{ego_neighbourhood_size}{The number of steps from the source node to
include nodes for each ego-network. NetEmd was proposed for individual nodes alone, hence the default value is 0.}
}
\value{
NetEMD measure for the two sets of discrete histograms (or graphs). If
(\code{return_details = FALSE}) then a list with the following named elements is returned
\code{net_emd}: the NetEMD for the set of histogram pairs (or graphs), \code{min_emds}:
the minimal EMD for each pair of histograms, \code{min_offsets}: the associated
offsets giving the minimal EMD for each pair of histograms
}
\description{
Calculates the mean minimum Earth Mover's Distance (EMD) between 
two sets of network features, after individually normalising the distribution
of each feature so that they have unit mass and unit variance.
This is calculated as follows:
  1. Normalise each histogram to have unit mass and unit variance.
  2. Find the minimum EMD between each pair of histograms.
  3. Take the average minimum EMD across all histogram pairs.
}
\examples{
 require(igraph)
 goldstd_1 <- graph.lattice(c(8,8)) 
 goldstd_2 <- graph.lattice(c(44,44)) 
 net_emd_one_to_one(graph_1=goldstd_1,graph_2=goldstd_2,feature_type="orbit",max_graphlet_size=5)
 
 plot(goldstd_1,vertex.size=0.8,vertex.label=NA)
 plot(goldstd_2,vertex.size=0.8,vertex.label=NA)
}
