% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{net_emds_for_all_graphs}
\alias{net_emds_for_all_graphs}
\title{NetEMDs between all graph pairs using provided Graphlet-based Degree 
Distributions}
\usage{
net_emds_for_all_graphs(gdds, method = "optimise", step_size = NULL,
  smoothing_window_width = 0, return_details = FALSE,
  mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{gdds}{List containing sets of Graphlet-based Degree Distributions for 
all graphs being compared}

\item{method}{The method to use to find the minimum EMD across all potential 
offsets for each pair of histograms. Default is "optimise" to use
R's built-in \code{optimise} method to efficiently find the offset with the 
minimal EMD. However, this is not guaranteed to find the global minimum if 
multiple local minima EMDs exist. You can alternatively specify the 
"fixed_step" method, which will exhaustively evaluate the EMD between the 
histograms at overlapping offsets separated by a fixed step. The size of the 
fixed step is 1/2 the the minimum spacing between locations in either
histogram after normalising to unit variance}

\item{step_size}{Additional optional argument for "fixed_step" method allowing
user to specify their own minumum step size. Note that this step size is 
applied to the histograms after they have been normalised to unit variance}

\item{smoothing_window_width}{Width of "top-hat" smoothing window to apply to
"smear" point masses across a finite width in the real domain. Default is 0, 
which  results in no smoothing. Care should be taken to select a 
\code{smoothing_window_width} that is appropriate for the discrete domain 
(e.g.for the integer domain a width of 1 is the natural choice)}

\item{return_details}{Logical indicating whether to return the individual
minimal EMDs and associated offsets for all pairs of histograms}
}
\value{
NetEMD measures between all pairs of graphs for which GDDs 
were provided. Format of returned data depends on the \code{return_details}
parameter. If set to FALSE, a list is returned with the following named
elements:\code{net_emd}: a vector of NetEMDs for each pair of graphs, 
\code{comparison_specification}: a table containing the graph names and
indices within the input GDD list for each pair of graphs compared.
If \code{return_details} is set to FALSE, the list also contains the following
matrices for each graph pair: \code{min_emds}: the minimal EMD for each GDD 
used to compute the NetEMD, \code{min_offsets}: the associated offsets giving
the minimal EMD for each GDD
}
\description{
NetEMDs between all graph pairs using provided Graphlet-based Degree 
Distributions
}
