% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{netdis.plot}
\alias{netdis.plot}
\title{Netdis comparisons between one graph and many other graphs.}
\usage{
netdis.plot(
  netdislist,
  whatrow = c(1, 2)[2],
  clustering_method = "ward.D",
  main = "Nedis"
)
}
\arguments{
\item{netdislist}{Default output of \code{netdis_many_to_many}.}

\item{whatrow}{Selection of the row in \code{netdis_many_to_many$comp_spec} to be used for plotting.}

\item{clustering_method}{Clusterng mehod as allowed in the \code{pheatmap} function from the \code{pheatmap} package.}

\item{main}{Title of the plot.}
}
\value{
Provides a heatmap and dendrogram for the network comparisons via \code{pheatmap}.
}
\description{
Provides a heatmap and dendrogram for the network comparisons via \code{pheatmap}.
}
