% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis}
\alias{netdis}
\title{Netdis - for one graphlet size}
\usage{
netdis(centred_graphlet_counts1, centred_graphlet_counts2, graphlet_size)
}
\arguments{
\item{centred_graphlet_counts1}{Centred Graphlet Counts for graph 1}

\item{centred_graphlet_counts2}{Centred Graphlet Counts for graph 2}

\item{graphlet_size}{The size of graphlets to use for the Netdis calculation
(only counts for graphlets of the specified size will be used). The size of
a graphlet is the number of nodes it contains.}
}
\value{
Netdis statistic calculated using centred counts for graphlets of
the specified size
}
\description{
Calculate Netdis statistic between two graphs from their Centred Graphlet
Counts (generated using \code{netdis_centred_graphlet_counts}) for graphlets 
of size \code{graphlet_size}.
}
