% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_centred_graphlet_counts}
\alias{netdis_centred_graphlet_counts}
\title{netdis_centred_graphlet_counts}
\usage{
netdis_centred_graphlet_counts(graphlet_counts, exp_graphlet_counts,
  max_graphlet_size)
}
\arguments{
\item{graphlet_counts}{Ego network graphlet counts for a query graph}

\item{exp_graphlet_counts}{Pre-calculated expected counts for each graphlet 
type for each ego network.}

\item{max_graphlet_size}{max graphlet size to calculate centred counts for.}
}
\value{
graphlet_counts minus exp_graphlet_counts for graphlets up to size
max_graphlet_size.
}
\description{
Centre counts by subtracting expected graphlet counts from actual graphlet
counts.
}
