% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{netdis_const_expected_counts}
\alias{netdis_const_expected_counts}
\title{Create matrix of constant value to use as expected counts.}
\usage{
netdis_const_expected_counts(graphlet_counts, const)
}
\arguments{
\item{graphlet_counts}{Ego network graphlet counts matrix to create expected
counts for.}

\item{const}{Constant expected counts value to use.}
}
\value{
Counts of value const with same shape and names as graphlet_counts.
}
\description{
Create matrix of constant value to use as expected counts.
}
