% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_expected_counts}
\alias{netdis_expected_counts}
\title{netdis_expected_counts}
\usage{
netdis_expected_counts(graphlet_counts, density_breaks,
  density_binned_reference_counts, max_graphlet_size, scale_fn = NULL)
}
\arguments{
\item{graphlet_counts}{Matrix of graphlet and node counts (columns) for a
nummber of ego networks (rows).}

\item{density_breaks}{Density values defining bin edges.}

\item{density_binned_reference_counts}{Reference network graphlet counts for
each density bin.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes are counted.}

\item{scale_fn}{Optional function to scale calculated expected counts, taking
\code{graphlet_counts} and \code{max_graphlet_size} as arguments,
and returning a scale factor that the looked up
\code{density_binned_reference_counts} values will be multiplied by.}
}
\description{
Calculates expected graphlet counts for each ego network based on its density
and pre-calculated reference density bins and graphlet counts for each bin.
}
