% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_expected_graphlet_counts_ego}
\alias{netdis_expected_graphlet_counts_ego}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
netdis_expected_graphlet_counts_ego(graph, max_graphlet_size,
  neighbourhood_size, density_breaks, density_binned_reference_counts,
  min_ego_nodes = 3, min_ego_edges = 1, scale_fn = NULL)
}
\arguments{
\item{graph}{A connected, undirected, simple reference graph as an
\code{igraph} object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes are counted.}

\item{neighbourhood_size}{The number of steps from the source node to include
node in ego-network.}

\item{density_breaks}{Density values defining bin edges.}

\item{density_binned_reference_counts}{Reference network graphlet counts for
each density bin.}

\item{min_ego_nodes}{Filter ego networks which have fewer
than min_ego_nodes nodes}

\item{min_ego_edges}{Filter ego networks which have fewer
than min_ego_edges edges}

\item{scale_fn}{Optional function to scale calculated expected counts, taking
\code{graph} and \code{max_graphlet_size} as arguments, and returning a scale
factor that the looked up \code{density_binned_reference_counts} values will
be multiplied by.

TODO: Remove @export prior to publishing}
}
\description{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
Temporarily accessible during development.
}
