% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_many_to_many}
\alias{netdis_many_to_many}
\title{Netdis between all graph pairs}
\usage{
netdis_many_to_many(graphs, ref_graph, comparisons = "many-to-many",
  max_graphlet_size = 4, neighbourhood_size = 2, min_ego_nodes = 3,
  min_ego_edges = 1,
  binning_fn = purrr::partial(binned_densities_adaptive,
  min_counts_per_interval = 5, num_intervals = 100),
  bin_counts_fn = purrr::partial(density_binned_counts, agg_fn = mean,
  scale_fn = scale_graphlet_counts_ego),
  exp_counts_fn = purrr::partial(netdis_expected_graphlet_counts_per_ego,
  scale_counts_fn = count_graphlet_tuples))
}
\arguments{
\item{graphs}{Query graphs}

\item{ref_graph}{Reference graph}

\item{comparisons}{Which comparisons to perform between graphs.
Can be "many-to-many" (all pairwise combinations) or "one-to-many"
(compare first graph in graphs to all other graphs.)}

\item{max_graphlet_size}{Generate graphlets up to this size}

\item{neighbourhood_size}{Ego network neighbourhood size}

\item{min_ego_nodes}{Filter ego networks which have fewer
than min_ego_nodes nodes}

\item{min_ego_edges}{Filter ego networks which have fewer
than min_ego_edges edges}

\item{min_bin_count}{Minimum number of ego networks in each density bin}

\item{num_bins}{Number of density bins to generate}
}
\value{
Netdis statistics between graph_1 and graph_2 for graphlet sizes
up to and including max_graphlet_size
}
\description{
Netdis between all graph pairs
}
