% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_one_to_one}
\alias{netdis_one_to_one}
\title{Netdis between two graphs}
\usage{
netdis_one_to_one(graph_1, graph_2, ref_graph, max_graphlet_size = 4,
  neighbourhood_size = 2, min_ego_nodes = 3, min_ego_edges = 1,
  min_bin_count = 5, num_bins = 100)
}
\arguments{
\item{graph_1}{First query graph}

\item{graph_2}{Second query graph}

\item{ref_graph}{Reference graph}

\item{max_graphlet_size}{Generate graphlets up to this size}

\item{neighbourhood_size}{Ego network neighbourhood size}

\item{min_ego_nodes}{Filter ego networks which have fewer
than min_ego_nodes nodes}

\item{min_ego_edges}{Filter ego networks which have fewer
than min_ego_edges edges}

\item{min_bin_count}{Minimum number of ego networks in each density bin}

\item{num_bins}{Number of density bins to generate}
}
\value{
Netdis statistics between graph_1 and graph_2 for graphlet sizes
up to and including max_graphlet_size
}
\description{
Netdis between two graphs
}
