% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{netemd.plot}
\alias{netemd.plot}
\title{Heatmap of NetEmd many-to-many comparisons}
\usage{
netemd.plot(netemdlist, clustering_method = "ward.D", main = "NetEmd")
}
\arguments{
\item{clustering_method}{Clustering method as allowed in the \code{pheatmap} function from the \code{pheatmap} package.}

\item{main}{Title of the plot.}

\item{netdislist}{Default output of \code{netdis_many_to_many}.}

\item{whatrow}{Selection of the row in \code{netdis_many_to_many$comp_spec} to be used for plotting.}
}
\value{
Provides a heatmap and dendrogram for the network comparisons via \code{pheatmap}.
}
\description{
Provides a heatmap and dendrogram for the network comparisons via \code{pheatmap}.
}
