% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{normalise_dhist_mass}
\alias{normalise_dhist_mass}
\title{Normalise a discrete histogram to unit mass}
\usage{
normalise_dhist_mass(dhist)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}
}
\value{
A discrete histogram normalised to have mass 1
}
\description{
Normalises a discrete histogram to unit mass by dividing each mass by the
total of the non-normalised masses
}
