% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{normalise_histogram_mass}
\alias{normalise_histogram_mass}
\title{Normalise histogram to unit mass}
\usage{
normalise_histogram_mass(bin_masses)
}
\arguments{
\item{bin_masses}{Bin masses for histogram}
}
\value{
normalised_masses Bin masses normalised to sum to 1
}
\description{
Normalises histogram to unit mass by dividing each bin mass by the total of 
the non-normalised bin masses
}
