% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{normalise_histogram_variance}
\alias{normalise_histogram_variance}
\title{Normalise histogram to unit variance}
\usage{
normalise_histogram_variance(bin_masses, bin_centres)
}
\arguments{
\item{bin_masses}{Bin masses for histogram}

\item{bin_centres}{Bin centres for histogram}
}
\value{
Bin centres normalised to give a histogram of variance 1
}
\description{
Normalises histogram to unit variance by dividing each bin centre by the 
standard deviation of the hsitogram
}
