% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{orbit_key}
\alias{orbit_key}
\title{Orbit key}
\usage{
orbit_key(max_graphlet_size)
}
\arguments{
\item{max_graphlet_size}{Maximum number of nodes graphlets can contain}
}
\value{
Metadata list with the following named fields:
\itemize{
  \item \code{max_nodes}: Maximum number of nodes graphlets can contain
  \item \code{id}: ID of each graphlet in format On, where n is in range 0 to 
 num_orbits
  \item \code{node_count}: Number of nodes contained within each graphlet
}
}
\description{
Metdata about orbit groups.
}
