% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{orbit_to_graphlet_counts}
\alias{orbit_to_graphlet_counts}
\title{Orbit to graphlet counts}
\usage{
orbit_to_graphlet_counts(orbit_counts)
}
\arguments{
\item{orbit_counts}{ORCA-format matrix containing counts of each graphlet
orbit (columns) at each vertex in the graph (rows)}
}
\value{
An ORCA-style matrix containing counts of each graphlet (columns) at
each vertex in the graph (rows)
}
\description{
Converts graphlet orbit counts at each vertex to graphlet counts at each
vertex by summing over all orbits contained within each graphlet
}
