% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{orca_graph}
\alias{orca_graph}
\title{Make arbitrary igraph ORCA compatible}
\usage{
orca_graph(graph)
}
\arguments{
\item{graph}{Original igraph graph object}
}
\value{
graph Processed igraph graph object
}
\description{
Takes a igraph graph object and does the following to make the graph 
compatible with the ORCA fast orbit calculation package:
  1. Makes the graph undirected
  2. Removes loops (where both endpoints of an edge are the same vertex)
  3. Removes multiple edges (i.e. ensuring only one edge exists for each 
     pair of endpoints)
  4. Removes isolated vertices (i.e. vertices with no edges after the 
     previous alterations)
}
