% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{read_all_graphs_as_orca_graphs}
\alias{read_all_graphs_as_orca_graphs}
\title{Load all graphs in a directory, converting to ORCA compatible graphs}
\usage{
read_all_graphs_as_orca_graphs(source_dir, format = "ncol",
  pattern = ".txt")
}
\arguments{
\item{source_dir}{Path to graph directory}

\item{format}{Format of graph files}

\item{pattern}{Filename pattern to match graph files}
}
\value{
A list of ORCA compatible graphs, with element names corresponding 
to the names of the files each graph was loaded from
}
\description{
Loads graphs from all files matching the given pattern in the given directory
and converts them to graphs compatible with the ORCA fast orbit 
counting package by:
  1. Making the graph undirected
  2. Removing loops (where both endpoints of an edge are the same vertex)
  3. Removing multiple edges (i.e. ensuring only one edge exists for each 
     pair of endpoints)
  4. Removing isolated vertices (i.e. vertices with no edges after the 
     previous alterations)
  5. Relabelling vertices with an integer index from 1 to numVertices
}
