% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{read_orca_graph}
\alias{read_orca_graph}
\title{Load graph from file and process to be ORCA compatible}
\usage{
read_orca_graph(file, format = "ncol")
}
\arguments{
\item{file}{Path to graph file}

\item{format}{Format of graph file}
}
\value{
An ORCA compatible igraph graph object
}
\description{
Loads a graph from file as a connected, undirected, simple 
\code{igraph} object.
}
