% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{read_simple_graph}
\alias{read_simple_graph}
\title{Read a graph from file, simplifying as requested}
\usage{
read_simple_graph(
  file,
  format,
  as_undirected = TRUE,
  remove_loops = TRUE,
  remove_multiple = TRUE,
  remove_isolates = TRUE
)
}
\arguments{
\item{file}{Path to file containing graph data}

\item{format}{Format of graph data. All formats supported by
\code{igraph::read_graph} are supported.}

\item{as_undirected}{If TRUE make graph edges undirected}

\item{remove_loops}{If TRUE, remove edgeds that connect a vertex to itself}

\item{remove_multiple}{If TRUE remove multiple edges connencting the same
pair of vertices}

\item{remove_isolates}{If TRUE, remove vertices with no edges after the
previous alterations have been made}
}
\value{
A simplified igraph graph object
}
\description{
Reads graph data from file, constructing an a igraph graph object, making the
requested subset of the following simplifications in the following order:
  1. Makes the graph undirected
  2. Removes loops (where both endpoints of an edge are the same vertex)
  3. Removes multiple edges (i.e. ensuring only one edge exists for each
     pair of endpoints)
  4. Removes isolated vertices (i.e. vertices with no edges after the
     previous alterations)
}
