% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{shift_dhist}
\alias{shift_dhist}
\title{Shift discrete histogram}
\usage{
shift_dhist(dhist, shift)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}

\item{shift}{The distance to add to all locations}
}
\value{
A shifted discrete histogram as a \code{dhist} object
}
\description{
Shift the locations of a discrete histogram rightwards on the x-axis by the
specified amount
}
