% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{shift_histogram}
\alias{shift_histogram}
\title{Shift histogram}
\usage{
shift_histogram(histogram, shift)
}
\arguments{
\item{histogram}{A histogram as a list with named members `masses` and `locations`}

\item{shift}{The distance to add to all bin locations}
}
\value{
A  shifted histogram as a list with named members `masses` and `locations`
}
\description{
Shift the bin locations of a histogram rightwards on the x-axis by the 
specificed shift
}
