% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{shift_to_next_alignment}
\alias{shift_to_next_alignment}
\title{Minimum shift to next alignment of two location vectors}
\usage{
shift_to_next_alignment(x1, x2, distance_matrix_prev = NULL,
  shift_prev = NULL)
}
\arguments{
\item{x1}{First location vector. This vector is being shifted rightwards}

\item{x2}{Second location vector. This vector is remaining unchanged.}
}
\value{
Minimum non-zero right-shift to apply to x1 to align at least one
element of x1 with at least one element of x2
}
\description{
Calculate minimum right shift of first location vector to make any pair of
locations from the two vectors equal
}
