% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{single_density_bin}
\alias{single_density_bin}
\title{For case where don't want to use binning, return a single bin which covers
the full range of possible density values (0 to 1).}
\usage{
single_density_bin(densities)
}
\arguments{
\item{densities}{Ego network density values (only used to return
a list of indexes of the required length.)}
}
\description{
For case where don't want to use binning, return a single bin which covers
the full range of possible density values (0 to 1).
}
