% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{sort_dhist}
\alias{sort_dhist}
\title{Sort discrete histogram}
\usage{
sort_dhist(dhist, decreasing = FALSE)
}
\arguments{
\item{dhist}{A discrete histogram as a \code{dhist} object}

\item{decreasing}{Logical indicating whether histograms should be sorted in 
increasing (default) or decreasing order of location}
}
\description{
Sort a discrete histogram so that locations are in increasing (default) or 
decreasing order
}
