% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{virusppi}
\alias{virusppi}
\title{Protein-protein interaction (PPI) networks for 5 microorganisms}
\format{
A list of \code{igraph} objects.
}
\source{
\strong{PPI data (EBV, HSV-1, KSHV, VZV):} Fossum E, Friedel CC,
Rajagopala SV, Titz B, Baiker A, Schmidt T, et al. (2009) Evolutionarily
Conserved Herpesviral Protein Interaction Networks. PLoS Pathog 5(9):
e1000570. \url{https://doi.org/10.1371/journal.ppat.1000570}. Data from Table
S2 in the supporting information.

\strong{PPI data (ECL):} Peregrín-Alvarez JM, Xiong X, Su C,
Parkinson J (2009) The Modular Organization of Protein Interactions in
Escherichia coli. PLoS Comput Biol 5(10): e1000523.
\url{https://doi.org/10.1371/journal.pcbi.1000523}

\strong{Taxonomy ground truth:} NCBI taxonomy database.
\url{https://www.ncbi.nlm.nih.gov/taxonomy}
}
\usage{
virusppi
}
\description{
A dataset containing the protein-protein interaction networks for the
following 5 microorganisms
\itemize{
 \item EBV
   \itemize{
     \item Common name: Epstein Barr virus
     \item Scientific name: Human gammaherpesvirus 4
     \item TaxonomyID: 10376
   }
 \item ECL
   \itemize{
     \item Common name: E.coli
     \item Scientific name: Escherichia coli
     \item TaxonomyID: 562
   }
 \item HSV-1
   \itemize{
     \item Common name: Herpes simplex virus type 1
     \item Scientific name: Human alphaherpesvirus 1
     \item TaxonomyID: 10298
   }
 \item KSHV
   \itemize{
     \item Common name: Karposi's Sarcoma-Associated Herpesvirus
     \item Scientific name: Human gammaherpesvirus 8
     \item TaxonomyID: 37296
   }
 \item VZV
   \itemize{
     \item Common name: Varicella zonster virus
     \item Scientific name: Human alphaherpesvirus 3
     \item TaxonomyID: 10335
   }
}
}
\keyword{datasets}
