% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{wtnets}
\alias{wtnets}
\title{World trade networks from 1962–2014}
\format{A list of \code{igraph} objects.}
\source{
\strong{World trade networks:} Feenstra RC,Lipsey RE, Deng H, Ma AC, and Mo H. (2005) World trade flows: 1962-2000. Technical report, National Bureau of Economic Research
}
\usage{
wtnets
}
\description{
The world trade data set consists of bilateral trade flows between countries for the years 1962–2014 [Feenstra et al., 2005, Division, 2015].
}
\keyword{datasets}
