% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_binning.R
\name{binned_densities_adaptive}
\alias{binned_densities_adaptive}
\title{binned_densities_adaptive}
\usage{
binned_densities_adaptive(densities, min_counts_per_interval,
  num_intervals)
}
\arguments{
\item{densities}{Density values to use for binning.}

\item{min_counts_per_interval}{Minimum count for each bin.}

\item{num_intervals}{Initial number of density bins to generate.
TODO: Remove @export prior to publishing}
}
\description{
Adaptive binning function guaranteeing a minimum number of entries in each
bin.
}
