% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{cost_matrix}
\alias{cost_matrix}
\title{Inter-bin cost matrix from bin centres}
\usage{
cost_matrix(bin_centres1, bin_centres2)
}
\arguments{
\item{bin_centres1}{Bin centres for histogram 1}

\item{bin_centres2}{Bin centres for histogram 2}
}
\description{
Generates a matrix for the cost of moving a unit of mass between each bin in 
histogram 1 and each bin in histogram 2.
}
