% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{count_graphlet_tuples}
\alias{count_graphlet_tuples}
\title{For each graphlet calculate the number of possible sets of k nodes in the
query graph, where k is the number of nodes in the graphlet.}
\usage{
count_graphlet_tuples(graph_graphlet_counts, max_graphlet_size)
}
\arguments{
\item{graph_graphlet_counts}{Node and graphlet counts for a single graph.}

\item{max_graphlet_size}{Determines the maximum size of graphlets included
in the tuple counts. Currently only size 4 and 5 are supported.}
}
\description{
For each graphlet calculate the number of possible sets of k nodes in the
query graph, where k is the number of nodes in the graphlet.
}
