% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{count_graphlet_tuples}
\alias{count_graphlet_tuples}
\title{FLAGUSED
For each graphlet calculate the number of possible sets of k nodes in the
query graph, where k is the number of nodes in the graphlet.}
\usage{
count_graphlet_tuples(graph, max_graphlet_size)
}
\arguments{
\item{graph}{A connected, undirected, simple graph as an \code{igraph}
object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets included
in the tuple counts.}
}
\description{
FLAGUSED
For each graphlet calculate the number of possible sets of k nodes in the
query graph, where k is the number of nodes in the graphlet.
}
