% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{count_graphlet_tuples_ego}
\alias{count_graphlet_tuples_ego}
\title{Run count_graphlet_tuples across pre-computed ego networks.}
\usage{
count_graphlet_tuples_ego(graphlet_counts, max_graphlet_size)
}
\arguments{
\item{graphlet_counts}{Matrix of graphlet and node counts (columns) for a
number of ego networks (rows).}

\item{max_graphlet_size}{Determines the maximum size of graphlets included
in the tuple counts. Currently only size 4 and 5 are supported.}
}
\description{
Run count_graphlet_tuples across pre-computed ego networks.
}
