% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{count_graphlets_ego_scaled}
\alias{count_graphlets_ego_scaled}
\title{Scaled graphlet count for ego-networks}
\usage{
count_graphlets_ego_scaled(graph, max_graphlet_size, neighbourhood_size)
}
\arguments{
\item{graph}{A connected, undirected, simple graph as an \code{igraph} object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
ORCA-format matrix containing counts of each graphlet (columns) at 
each vertex in the graph (rows).
}
\description{
Scales the ego-network count for each graphlet by dividing it by the total
number of possible groupings of nodes in the ego-network with the same number
of nodes as the graphlet.
}
