% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{count_orbits}
\alias{count_orbits}
\title{Calculate graphlet orbit counts}
\usage{
count_orbits(graph, max_graphlet_size)
}
\arguments{
\item{graph}{A connected, undirected, simple graph as an \code{igraph} object.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
ORCA-format matrix containing counts of each graphlet
orbit (columns) at each vertex in the graph (rows).
}
\description{
Calculates graphlet orbit counts from \code{igraph} graph object,
using the ORCA fast graphlet orbit counting package.
}
