% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{cross_comparison_spec}
\alias{cross_comparison_spec}
\title{Generate a cross-comparison specification}
\usage{
cross_comparison_spec(named_list)
}
\arguments{
\item{named_list}{A named list of items for which an exhaustive pair-wise
cross-comparison is required.}
}
\value{
A matrix with one row for each possible pair-wise combination
of elements from the provided named list. The first and second columns
contain the names of the elements in the pair and the third and fourth
columns contain the indexes of these elements in the provided list.
}
\description{
Creates a cross-comparison matrix with all possible pair-wise combinations
of elements from the provided list.
}
