% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{density_binned_counts}
\alias{density_binned_counts}
\title{Used to calculate aggregated graphlet counts for each density bin.}
\usage{
density_binned_counts(
  graphlet_counts,
  density_interval_indexes,
  agg_fn = mean,
  scale_fn = NULL,
  max_graphlet_size = NULL
)
}
\arguments{
\item{graphlet_counts}{Graphlet and node counts (columns) for a number of
ego_networks (rows).}

\item{density_interval_indexes}{Density bin index for
each ego network.}

\item{agg_fn}{Function to aggregate counts in each bin
(default \code{agg_fn = mean}).}

\item{scale_fn}{Optional function to apply a transformation/scaling
to the raw graphlet_counts. Must have arguments \code{graphlet_counts} and
\code{max_graphlet_size}, and return a transformed \code{graphlet_counts}
object with the same number of rows as the input, and columns for all
graphlets up to \code{max_graphlet_size}.}

\item{max_graphlet_size}{Optionally passed and used by scale_fn.}
}
\description{
Used to calculate aggregated graphlet counts for each density bin.
}
