% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{density_binned_counts}
\alias{density_binned_counts}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
density_binned_counts(graphlet_counts, density_interval_indexes,
  agg_fn = mean, scale_fn = NULL, ego_networks = NULL,
  max_graphlet_size = NULL)
}
\arguments{
\item{graphlet_counts}{Graphlet counts for a number of ego_networks.}

\item{density_interval_indexes}{Density bin index for 
each ego network.}

\item{agg_fn}{Function to aggregate counts in each bin
(default \code{agg_fn = mean}).}

\item{scale_fn}{Optional function to apply a transformation
to graphlet_counts, must have arguments graphlet_counts,
ego_networks and max_graphlet_size.}

\item{ego_networks}{Optionally passed and used by scale_fn.}

\item{max_graphlet_size}{Optionally passed and used by scale_fn.}
}
\description{
Used to calculate expected graphlet counts for each density bin.
}
