% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhist.R
\name{dhist}
\alias{dhist}
\title{Discrete histogram constructor}
\usage{
dhist(locations, masses)
}
\arguments{
\item{\code{locations}}{A 1D numeric vector specifying the discrete locations
of the histogram bins}

\item{\code{masses}}{A 1D numeric vector specifying the mass present at each 
location}
}
\value{
A sparse discrete histogram. Format is a \code{dhist} object, which
is a list of class \code{dhist} with the following named elements:
\itemize{
  \item \code{locations}: A 1D numeric vector of discrete locations
  \item \code{masses}: A 1D numeric vector of the mass present at each location
}
Note that locations where no mass is present are not included in the returned
\code{dhist} object. Mass in these discrete histograms is treated as being
present precisely at the specified location. Discrete histograms should not be used
for data where observations have been grouped into bins representing ranges 
of observation values.
}
\description{
Creates a discrete histogram object of class \code{dhist}, with bin 
\code{locations} and \code{masses} set to the 1D numeric vectors provided.
}
