% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd}
\alias{emd}
\title{Earth Mover's Distance (EMD)}
\usage{
emd(histogram1, histogram2)
}
\arguments{
\item{histogram1}{Histogram 1 as list with names members `masses` and `locations`}

\item{histogram2}{Histogram 2 as list with names members `masses` and `locations`}
}
\value{
Earth Mover's Distance between the two input histograms
}
\description{
Takes two sets of histogram bin masses and bin centres and calculates the 
Earth Mover's Distance between the two histograms
}
