% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd}
\alias{emd}
\title{Earth Mover's Distance (EMD)}
\usage{
emd(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A \code{dhist} discrete histogram object}

\item{dhist2}{A \code{dhist} discrete histogram object}
}
\value{
Earth Mover's Distance between the two discrete histograms
}
\description{
Calculates the Earth Mover's Distance (EMD) between two discrete histograms
}
