% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd_cs}
\alias{emd_cs}
\title{Earth Mover's Distance (EMD) using the difference of cumulative sums method}
\usage{
emd_cs(dhist1, dhist2)
}
\arguments{
\item{dhist1}{A discrete histogram as a \code{dhist} object}

\item{dhist2}{A discrete histogram as a \code{dhist} object}
}
\value{
Earth Mover's Distance between the two input histograms
}
\description{
Takes two discrete histograms and calculates the Wasserstein / Earth Mover's
Distance between the two histograms by summing the absolute difference 
between the two cumulative histograms.
}
\references{
Calculation of the Wasserstein Distance Between Probability Distributions on the Line
S. S. Vallender, Theory of Probability & Its Applications 1974 18:4, 784-786
\url{http://dx.doi.org/10.1137/1118101}
}
