% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd_cs}
\alias{emd_cs}
\title{Earth Mover's Distance (EMD) using the difference of cumulative sums method}
\usage{
emd_cs(bin_masses1, bin_masses2, bin_centres1, bin_centres2)
}
\arguments{
\item{bin_masses1}{Bin masses for histogram 1}

\item{bin_masses2}{Bin masses for histogram 2}

\item{bin_centres1}{Bin centres for histogram 1}

\item{bin_centres2}{Bin centres for histogram 2}
}
\description{
Takes two sets of histogram bin masses and bin centres and calculates the 
Earth Mover's Distance between the two histograms by summing the absolute 
difference between the two cumulative histograms.
}
\references{
Calculation of the Wasserstein Distance Between Probability Distributions on the Line
S. S. Vallender, Theory of Probability & Its Applications 1974 18:4, 784-786
\url{http://dx.doi.org/10.1137/1118101}
}
