% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{emd_lp}
\alias{emd_lp}
\title{Earth Mover's Distance (EMD) using linear programming (LP)}
\usage{
emd_lp(bin_masses1, bin_masses2, bin_centres1, bin_centres2)
}
\arguments{
\item{bin_masses1}{Bin masses for histogram 1}

\item{bin_masses2}{Bin masses for histogram 2}

\item{bin_centres1}{Bin centres for histogram 1}

\item{bin_centres2}{Bin centres for histogram 2}
}
\description{
Takes two sets of histogram bin masses and bin centres and calculates the 
Earth Mover's Distance between the two histograms by solving the Transport
Problem using linear programming.
}
