% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{gdd}
\alias{gdd}
\title{Graphlet-based degree distributions (GDDs)}
\usage{
gdd(graph, feature_type = "orbit", max_graphlet_size = 4)
}
\arguments{
\item{feature_type}{Type of graphlet-based feature to count: "graphlet"
counts the number of graphlets each node participates in; "orbit" calculates
the number of graphlet orbits each node participates in.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}

\item{indexed_edges}{A 2 x numEdges edgelist with vertices labelled with 
integer indices, with an optional "vertex_names" attribute}
}
\value{
List of graphlet-based degree distributions, with each distribution
represented as a \code{dhist} discrete histogram object.
}
\description{
Generates graphlet-based degree distributions from indexed ege lists,
using the ORCA fast graphlet orbit counting package.
}
