% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{gdd}
\alias{gdd}
\title{Graphlet-based degree distributions (GDDs)}
\usage{
gdd(graph, feature_type = "orbit", max_graphlet_size = 4,
  ego_neighbourhood_size = 0)
}
\arguments{
\item{graph}{An \code{igraph} graph object. The graph must have the following
properties:' 
  1. Be undirected
  2. Have no self-loops (where both endpoints of an edge are the same vertex)
  3. Not have multiple edges multiple edges (i.e. at most one edge exists for
     each pair of vertices)
  4. No isolated vertices (i.e. vertices with no edges)}

\item{feature_type}{Type of graphlet-based feature to count: "graphlet"
counts the number of graphlets each node participates in; "orbit" calculates
the number of graphlet orbits each node participates in.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count. 
Only graphlets containing up to \code{max_graphlet_size} nodes will be counted.}
}
\value{
List of graphlet-based degree distributions, with each distribution
represented as a \code{dhist} discrete histogram object.
}
\description{
Generates graphlet-based degree distributions from \code{igraph} graph object,
using the ORCA fast graphlet orbit counting package.
}
