% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{godd}
\alias{godd}
\title{Graphlet Orbit Degree Distributions}
\usage{
godd(indexed_edges)
}
\arguments{
\item{indexed_edges}{A 2 x numEdges edgelist with vertices labelled with 
integer indices, with an optional "vertex_names" attribute}
}
\value{
Graphlet degree histograms: List of degree histograms for each 
graphlet orbit, with each histogram represented as a \code{dhist} discrete
histogram object
}
\description{
Calculates the set of graphlet orbit degree distributions for graphlets of
up to 4 nodes.
}
