% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{graph_features_to_histograms}
\alias{graph_features_to_histograms}
\title{Convert a matrix of node level features to a "discrete histogram" for
each feature.}
\usage{
graph_features_to_histograms(features_matrix)
}
\arguments{
\item{features_matrix}{A matrix whose rows represent nodes and whose columns
represent different node level features. This means that entry ij provides
the value of feature j for node i.}
}
\value{
Feature histograms: List of "discrete histograms" for each feature
}
\description{
Converts a matrix of node level features (e.g. for example counts
of multiple graphlets or orbits at each node) to
a set of histogram like objects (observed frequency distribution of each
feature/column)
}
