% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{make_named_ego_graph}
\alias{make_named_ego_graph}
\title{FLAGUSED
Get ego-networks for a graph as a named list}
\usage{
make_named_ego_graph(graph, order, min_ego_nodes = 3,
  min_ego_edges = 1, ...)
}
\arguments{
\item{graph}{An \code{igraph} object}

\item{order}{The number of steps from the source node to include
nodes for each ego-network.}

\item{min_ego_nodes}{Only ego networks with at least \code{min_ego_nodes}
nodes are returned.}

\item{min_ego_edges}{Only ego networks with at least \code{min_ego_edges}
edges are returned.}

\item{...}{Additional parameters to be passed to the underlying
\code{igraph::make_ego_graph} function used.}
}
\description{
Simple wrapper for the \code{igraph::make_ego_graph} function that names
each ego-network in the returned list with the name of the node in the
original graph that the ego-network was generated from
}
