% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{mean_density_binned_graphlet_counts}
\alias{mean_density_binned_graphlet_counts}
\title{mean_density_binned_graphlet_counts}
\usage{
mean_density_binned_graphlet_counts(
  graphlet_counts,
  density_interval_indexes,
  agg_fn = mean
)
}
\arguments{
\item{graphlet_counts}{Graphlet counts for a number of ego_networks.}

\item{density_interval_indexes}{Density bin index for
each ego network in graphlet_counts.}

\item{agg_fn}{Function to aggregate counts in each bin
(default \code{agg_fn = mean}).}
}
\description{
Calculate mean (dy default) graphlet counts for ego networks in each density
bin.
}
