% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_emd.R
\name{net_emd}
\alias{net_emd}
\title{NetEMD Network Earth Mover's Distance}
\usage{
net_emd(bin_masses1, bin_masses2, bin_centres1, bin_centres2, step = 1)
}
\arguments{
\item{bin_masses1}{Bin masses for histogram 1}

\item{bin_masses2}{Bin masses for histogram 2}

\item{bin_centres1}{Bin centres for histogram 1}

\item{bin_centres2}{Bin centres for histogram 2}
}
\value{
NetEMD
}
\description{
Calculates the minimum Earth Mover's Distance (EMD) between two histograms.
This is calculated as follows:
  1. Normalise each histogram to have unit mass and unit variance
  2. "Slide" histogram 1 over histogram 2 by varying the bin offset between 
     the two histograms, calculating the EMD at each offset.
  3. The NetEMD is the minimum EMD observed across all offsets
}
