% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_centred_graphlet_counts_ego}
\alias{netdis_centred_graphlet_counts_ego}
\title{TODO: Remove @export prior to publishing}
\usage{
netdis_centred_graphlet_counts_ego(graph, max_graphlet_size,
  neighbourhood_size, expected_ego_count_fn = NULL, min_ego_nodes = 3,
  min_ego_edges = 1)
}
\description{
TODO: Remove @export prior to publishing
}
