% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_expected_graphlet_counts_ego}
\alias{netdis_expected_graphlet_counts_ego}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
netdis_expected_graphlet_counts_ego(graph, max_graphlet_size,
  neighbourhood_size, density_breaks, density_binned_reference_counts,
  min_ego_nodes = 3, min_ego_edges = 1)
}
\description{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to 
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
Temporarily accessible during development. 
TODO: Remove @export prior to publishing
}
