% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_expected_graphlet_counts_per_ego}
\alias{netdis_expected_graphlet_counts_per_ego}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
netdis_expected_graphlet_counts_per_ego(graphlet_counts, density_breaks,
  density_binned_reference_counts, max_graphlet_size, scale_fn = NULL)
}
\arguments{
\item{density_breaks}{Density values defining bin edges.}

\item{density_binned_reference_counts}{Reference network graphlet counts for
each density bin.}

\item{max_graphlet_size}{Determines the maximum size of graphlets to count.
Only graphlets containing up to \code{max_graphlet_size} nodes are counted.}

\item{scale_fn}{Optional function to scale calculated expected counts, taking
\code{graph} and \code{max_graphlet_size} as arguments, and returning a scale
factor that the looked up \code{density_binned_reference_counts} values will
be multiplied by.

#' Temporarily accessible during development.
TODO: Remove @export prior to publishing}

\item{ego_networks}{The number of steps from the source node to include
node in ego-network.}
}
\description{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
}
