% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_expected_graphlet_counts_per_ego}
\alias{netdis_expected_graphlet_counts_per_ego}
\title{INTERNAL FUNCTION - Do not call directly}
\usage{
netdis_expected_graphlet_counts_per_ego(ego_networks, density_breaks,
  density_binned_reference_counts, max_graphlet_size, scale_fn = NULL)
}
\description{
JACK To follow through logic of paper steps, wanted to pass
ego networks to the function, not the input query graph
(as in netdis_expected_graphlet_counts_ego above).
}
\details{
Used by \code{netdis_expected_graphlet_counts_ego_fn} to
generate a function for calculating expected ego-network graphlet counts
from the statistics of a provided reference graph.
Temporarily accessible during development.
TODO: Remove @export prior to publishing
}
