% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_many_to_many}
\alias{netdis_many_to_many}
\title{Compute any of the Netdis variants between all graph pairs.}
\usage{
netdis_many_to_many(
  graphs = NULL,
  ref_graph = NULL,
  comparisons = "many-to-many",
  max_graphlet_size = 4,
  neighbourhood_size = 2,
  min_ego_nodes = 3,
  min_ego_edges = 1,
  binning_fn = NULL,
  bin_counts_fn = NULL,
  exp_counts_fn = NULL,
  graphlet_counts = NULL,
  graphlet_counts_ref = NULL
)
}
\arguments{
\item{graphs}{A named list of simplified igraph graph objects (undirected
graphs excluding loops, multiple edges), such as those
obtained by using \code{read_simple_graphs}.}

\item{ref_graph}{Controls how expected counts are calculated. Either:
1) A numeric value - used as a constant expected counts value for all query
graphs.
2) A simplified \code{igraph} object - used as a reference graph from which
expected counts are calculated for all query graphs.
3) NULL (default) - Expected counts will be calculated based on the properties of the
query graphs themselves. (Geometric-Poisson approximation).}

\item{comparisons}{Which comparisons to perform between graphs.
Can be "many-to-many" (all pairwise combinations) or "one-to-many"
(compare first graph in graphs to all other graphs.)}

\item{max_graphlet_size}{Generate graphlets up to this size. Currently only 4 (default) and 5 are supported.}

\item{neighbourhood_size}{Ego network neighbourhood size (default 2).}

\item{min_ego_nodes}{Filter ego networks which have fewer
than min_ego_nodes nodes (default 3).}

\item{min_ego_edges}{Filter ego networks which have fewer
than min_ego_edges edges (default 1).}

\item{binning_fn}{Function used to bin ego network densities. Takes edge \code{densities}
as its single argument, and returns a named list including, the input \code{densities}, the resulting bin \code{breaks} (vector of density bin limits), and the vector \code{interval_indexes} which states to what bin each of the individual elements in \code{densities} belongs to.
ego network). If \code{NULL}, then the method \code{binned_densities_adaptive} with
\code{min_counts_per_interval = 5} and \code{num_intervals = 100} is used (default: NULL).}

\item{bin_counts_fn}{Function used to calculate expected graphlet counts in
each density bin. Takes \code{graphlet_counts}, \code{interval_indexes}
(bin indexes) and \code{max_graphlet_size} as arguments. If \code{bin_counts_fn} is \code{NULL}, (default),
it will apply either the approach from the original Netdis paper, or the respective Geometric-Poisson
approximation; depending on the values of \code{ref_graph} and \code{graphlet_counts_ref}.}

\item{exp_counts_fn}{Function used to map from binned reference counts to
expected counts for each graphlet in each ego network of the query graphs.
Takes \code{ego_networks}, \code{density_bin_breaks},
\code{binned_graphlet_counts}, and \code{max_graphlet_size} as arguments.
If \code{exp_counts_fn} is \code{NULL}, (default), it will apply
either the approach from the original Netdis paper, or the respective Geometric-Poisson approximation; depending on the
values of \code{ref_graph} and \code{graphlet_counts_ref}.}

\item{graphlet_counts}{Pre-generated graphlet counts (default: NULL). If the
\code{graphlet_counts} argument is defined then \code{graphs} will not be
used.
A named list of matrices containing counts of each graphlet (columns) for
each ego-network (rows) in the input graph. Columns are labelled with
graphlet IDs and rows are labelled with the ID of the central node in each
ego-network. As well as graphlet counts, each matrix  must contain an
additional column labelled "N" including the node count for
each ego network.}

\item{graphlet_counts_ref}{Pre-generated reference graphlet counts (default: NULL). Matrix containing counts 
of each graphlet (columns) for each ego-network (rows) in the input graph. Columns are labelled with
graphlet IDs and rows are labelled with the ID of the central node in each
ego-network. As well as graphlet counts, each matrix  must contain an
additional column labelled "N" including the node count for
each ego network. 
If the \code{graphlet_counts_ref} argument is defined then \code{ref_graph} will not
be used.}
}
\value{
Netdis statistics between query graphs for graphlet sizes
up to and including max_graphlet_size.
}
\description{
Compute any of the Netdis variants between all graph pairs.
}
