% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{netdis_one_to_one}
\alias{netdis_one_to_one}
\title{Netdis between two graphs}
\usage{
netdis_one_to_one(
  graph_1 = NULL,
  graph_2 = NULL,
  ref_graph = 0,
  max_graphlet_size = 4,
  neighbourhood_size = 2,
  min_ego_nodes = 3,
  min_ego_edges = 1,
  binning_fn = purrr::partial(binned_densities_adaptive, min_counts_per_interval = 5,
    num_intervals = 100),
  bin_counts_fn = purrr::partial(density_binned_counts, agg_fn = mean, scale_fn =
    scale_graphlet_counts_ego),
  exp_counts_fn = purrr::partial(netdis_expected_counts, scale_fn =
    count_graphlet_tuples),
  graphlet_counts_1 = NULL,
  graphlet_counts_2 = NULL
)
}
\arguments{
\item{graph_1}{A simplified igraph graph object.}

\item{graph_2}{A simplified igraph graph object.}

\item{ref_graph}{Controls how expected counts are calculated. Either:
1) A numeric value - used as a constant expected counts value for all query
graphs (DEFAULT: 0).
2) A simplified \code{igraph} object - used as a reference graph from which
expected counts are calculated for all query graphs.
3) NULL - Expected counts will be calculated based on the properties of the
query graphs themselves.}

\item{max_graphlet_size}{Generate graphlets up to this size.}

\item{neighbourhood_size}{Ego network neighbourhood size.}

\item{min_ego_nodes}{Filter ego networks which have fewer
than min_ego_nodes nodes.}

\item{min_ego_edges}{Filter ego networks which have fewer
than min_ego_edges edges.}

\item{binning_fn}{Function used to bin ego network densities. Takes densities
as its single argument, and returns a named list including keys \code{breaks}
(list of bin edges) and \code{interval_indexes} (density bin index for each
ego network). (Default: \code{binned_densities_adaptive} with
\code{min_counts_per_interval = 5} and \code{num_intervals = 100}).}

\item{bin_counts_fn}{Function used to calculate expected graphlet counts in
each density bin. Takes \code{graphlet_counts}, \code{interval_indexes}
(bin indexes) and \code{max_graphlet_size} as arguments.
(Default: \code{density_binned_counts} with \code{agg_fn = mean} and
\code{scale_fn = scale_graphlet_counts_ego}, which mirrors the
approach used in the original netdis paper).}

\item{exp_counts_fn}{Function used to map from binned reference counts to
expected counts for each graphlet in each ego network of the query graphs.
Takes \code{ego_networks}, \code{density_bin_breaks},
\code{binned_graphlet_counts}, and \code{max_graphlet_size} as arguments.
(Default: \code{netdis_expected_counts} with
\code{scale_fn = count_graphlet_tuples}, which mirrors the approach used in
the original netdis paper).}

\item{graphlet_counts_1}{Pre-generated graphlet counts for the first query
graph. If the \code{graphlet_counts_1} argument is defined then
\code{graph_1} will not be used.}

\item{graphlet_counts_2}{Pre-generated graphlet counts for the second query
graph. If the \code{graphlet_counts_2} argument is defined then
\code{graph_2} will not be used.}
}
\value{
Netdis statistics between graph_1 and graph_2 for graphlet sizes
up to and including max_graphlet_size
}
\description{
Netdis between two graphs
}
