% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{read_all_graphs_as_orca_graphs}
\alias{read_all_graphs_as_orca_graphs}
\title{Load all graphs in a directory, converting to ORCA compatible graphs}
\usage{
read_all_graphs_as_orca_graphs(source_dir, format = "ncol",
  pattern = ".txt")
}
\arguments{
\item{source_dir}{Path to graph directory}

\item{format}{Format of graph files}

\item{pattern}{Filename pattern to match graph files}
}
\value{
A list of ORCA compatible graphs, with element names corresponding 
to the names of the files each graph was loaded from
}
\description{
Loads graphs from all files matching the given pattern in the given directory
as connected, undirected, simple \code{igraph} objects.
}
