% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orca_interface.R
\name{read_orca_graph}
\alias{read_orca_graph}
\title{Load graph from file and process to be ORCA compatible}
\usage{
read_orca_graph(file, format = "ncol")
}
\arguments{
\item{file}{Path to graph file}

\item{format}{Format of graph file}
}
\value{
An ORCA compatible igraph graph object
}
\description{
Loads a graph from file into an igraph graph object and does the following 
to make the graph compatible with the ORCA fast orbit calculation package:
  1. Makes the graph undirected
  2. Removes loops (where both endpoints of an edge are the same vertex)
  3. Removes multiple edges (i.e. ensuring only one edge exists for each 
     pair of endpoints)
  4. Removes isolated vertices (i.e. vertices with no edges after the 
     previous alterations)
}
