% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{scale_graphlet_count}
\alias{scale_graphlet_count}
\title{Divide graphlet counts by pre-computed scaling factor from
\code{count_graphlet_tuples} output.}
\usage{
scale_graphlet_count(graphlet_count, graphlet_tuples)
}
\arguments{
\item{graphlet_count}{Pre-computed graphlet counts.}

\item{graphlet_tuples}{Pre-computed \code{count_graphlet_tuples} output.}
}
\description{
Divide graphlet counts by pre-computed scaling factor from
\code{count_graphlet_tuples} output.
}
