% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_netdis.R
\name{scale_graphlet_counts_ego}
\alias{scale_graphlet_counts_ego}
\title{Scale graphlet counts for an ego network by the n choose k possible
choices of k nodes in that ego-network, where n is the number of nodes
in the ego network and k is the number of nodes in the graphlet.}
\usage{
scale_graphlet_counts_ego(graphlet_counts, max_graphlet_size)
}
\arguments{
\item{graphlet_counts}{Pre-calculated graphlet counts for each ego_network.}

\item{max_graphlet_size}{Determines the maximum size of graphlets included
in graphlet_counts. Currently only size 4 and 5 are supported.}
}
\value{
scaled graphlet counts.
}
\description{
Scale graphlet counts for an ego network by the n choose k possible
choices of k nodes in that ego-network, where n is the number of nodes
in the ego network and k is the number of nodes in the graphlet.
}
