% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{single_density_bin}
\alias{single_density_bin}
\title{For case where don't want to use binning, return
a single bin which covers full range of possible
densities.}
\usage{
single_density_bin(densities)
}
\description{
For case where don't want to use binning, return
a single bin which covers full range of possible
densities.
}
