% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_net_dis.R
\name{zeros_to_ones}
\alias{zeros_to_ones}
\title{Replace zero values in a vector with ones. Used by
\code{scale_graphlet_count} to prevent divide by
zero errors.}
\usage{
zeros_to_ones(v)
}
\arguments{
\item{v}{A vector.
TODO remove export}
}
\description{
Replace zero values in a vector with ones. Used by
\code{scale_graphlet_count} to prevent divide by
zero errors.
}
